requirements.txt
discord.py
pymongo~=3.10.1
dnspython
